* add jwt-decode package. * add useAuth for create/update/delete restriction.

+++ src/pages/item/delete.js
@@ -2,6 +2,7 @@
 
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
+import useAuth from '../../utils/useAuth';
 
 const DeleteItem = () => {
   const base_url = process.env.REACT_APP_API_URL;
@@ -23,6 +24,7 @@ const DeleteItem = () => {
           image: jsonResp.singleItem.image,
           price: jsonResp.singleItem.price,
           description: jsonResp.singleItem.description,
+          email: jsonResp.singleItem.email,
         });
       } catch (error) {
         console.log(error);
@@ -48,18 +50,23 @@ const DeleteItem = () => {
       alert('アイテム削除失敗');
     }
   };
-  return (
-    <div>
-      <h1>アイテム削除</h1>
-      <form onSubmit={handleSubmit}>
-        <h2>{deleteItem.title}</h2>
-        {deleteItem.image && <img src={deleteItem.image} alt="item" />}
-        <h3>￥{Number(deleteItem.price).toLocaleString()}</h3>
-        <p>{deleteItem.description}</p>
-        <button>削除</button>
-      </form>
-    </div>
-  );
+  const loginUser = useAuth();
+  if (loginUser === deleteItem.email) {
+    return (
+      <div>
+        <h1>アイテム削除</h1>
+        <form onSubmit={handleSubmit}>
+          <h2>{deleteItem.title}</h2>
+          {deleteItem.image && <img src={deleteItem.image} alt="item" />}
+          <h3>￥{Number(deleteItem.price).toLocaleString()}</h3>
+          <p>{deleteItem.description}</p>
+          <button>削除</button>
+        </form>
+      </div>
+    );
+  } else {
+    return <h1>権限がありません</h1>;
+  }
 };
 
 export default DeleteItem;
